requirements.txt
flask==2.0.3
torch==2.0.1
transformers==4.30.0
pillow==9.5.0